Arrumar sequência de cenas (fase 2) Fixes #162
