+++ src/config.ts
@@ -18,12 +18,12 @@ const config: Types.Core.GameConfig = {
     gamepad: true,
   },
   scene: [
-    Scene0,
+    Scene5,
     Scene1,
     Scene2,
     Scene3,
     Scene4,
-    Scene5,
+    Scene0,
     Scene6,
     Scene7,
     Scene8,
